pyproject.toml
[tool.poetry]
name = "imagelib"
version = "0.1.1"
description = ""
authors = ["Mokashi,Chinmay <CMokashi@mdanderson.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
absl-py = "2.2.0"
annotated-types = "0.7.0"
antspynet = "0.2.9"
antspyx = "0.5.4"
appnope = "0.1.4"
asttokens = "3.0.0"
astunparse = "1.6.3"
attrs = "25.3.0"
bids-validator = "1.14.7.post0"
bidsschematools = "1.0.4"
certifi = "2025.1.31"
charset-normalizer = "3.4.1"
click = "8.1.8"
comm = "0.2.2"
contourpy = "1.3.1"
cycler = "0.12.1"
debugpy = "1.8.13"
decorator = "5.2.1"
docopt = "0.6.2"
executing = "2.2.0"
flatbuffers = "25.2.10"
fonttools = "4.56.0"
formulaic = "1.1.1"
frozendict = "2.4.6"
fsspec = "2025.3.0"
gast = "0.6.0"
google-pasta = "0.2.0"
grpcio = "1.71.0"
h5py = "3.13.0"
idna = "3.10"
interface-meta = "1.3.0"
ipykernel = "6.29.5"
ipython = "9.0.2"
ipython-pygments-lexers = "1.1.1"
ipywidgets = "8.1.5"
jedi = "0.19.2"
joblib = "1.4.2"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
jupyter-client = "8.6.3"
jupyter-core = "5.7.2"
jupyterlab-widgets = "3.0.13"
keras = "3.9.0"
kiwisolver = "1.4.8"
libclang = "18.1.1"
lxml = "5.3.1"
markdown = "3.7"
markdown-it-py = "3.0.0"
markupsafe = "3.0.2"
matplotlib = "3.10.1"
matplotlib-inline = "0.1.7"
mdurl = "0.1.2"
ml-dtypes = "0.4.1"
namex = "0.0.8"
nest-asyncio = "1.6.0"
nibabel = "5.3.2"
nilearn = "0.11.1"
num2words = "0.5.14"
numpy = "1.26.4"
opt-einsum = "3.4.0"
optree = "0.14.1"
packaging = "24.2"
pandas = "2.2.3"
parso = "0.8.4"
patsy = "1.0.1"
pexpect = "4.9.0"
pillow = "11.1.0"
platformdirs = "4.3.7"
prompt-toolkit = "3.0.50"
protobuf = "4.25.6"
psutil = "7.0.0"
ptyprocess = "0.7.0"
pure-eval = "0.2.3"
pybids = "0.19.0"
pydantic = "2.10.6"
pydantic-core = "2.27.2"
pydicom = "3.0.1"
pygments = "2.19.1"
pyparsing = "3.2.3"
python-dateutil = "2.9.0.post0"
pytz = "2025.2"
pyyaml = "6.0.2"
pyzmq = "26.3.0"
referencing = "0.36.2"
requests = "2.32.3"
rich = "13.9.4"
rpds-py = "0.23.1"
scikit-learn = "1.6.1"
scipy = "1.15.2"
setuptools = "78.1.0"
six = "1.17.0"
sqlalchemy = "2.0.39"
stack-data = "0.6.3"
statsmodels = "0.14.4"
tensorboard = "2.17.1"
tensorboard-data-server = "0.7.2"
tensorflow = "2.17.0"
termcolor = "2.5.0"
threadpoolctl = "3.6.0"
tornado = "6.4.2"
traitlets = "5.14.3"
typing-extensions = "4.12.2"
tzdata = "2025.2"
universal-pathlib = "0.2.6"
urllib3 = "2.3.0"
wcwidth = "0.2.13"
webcolors = "24.11.1"
werkzeug = "3.1.3"
wheel = "0.45.1"
widgetsnbextension = "4.0.13"
wrapt = "1.17.2"
aiohappyeyeballs = "2.6.1"
aiohttp = "3.11.14"
aiosignal = "1.3.2"
frozenlist = "1.5.0"
imageio = "2.37.0"
lazy-loader = "0.4"
more-itertools = "10.6.0"
msgpack = "1.1.0"
multidict = "6.2.0"
networkx = "3.4.2"
propcache = "0.3.0"
scikit-image = "0.25.2"
tifffile = "2025.3.13"
trame = "3.8.1"
trame-client = "3.6.1"
trame-server = "3.4.0"
trimesh = "4.6.5"
wslink = "2.3.3"
yarl = "1.18.3"
pooch = "1.8.2"
pyvista = "0.44.2"
scooby = "0.10.0"
vtk = "9.3.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
